+++ dtslint/ts4.1/Json.ts
@@ -1,0 +1,26 @@
+import * as _ from '../../src/Json'
+
+//
+// stringify
+//
+
+// // $ExpectError
+// _.stringify(undefined)
+// // $ExpectError
+// _.stringify(() => {})
+// // $ExpectError
+// _.stringify(Symbol())
+
+interface AB {
+  readonly a: string
+  readonly b: number
+}
+
+const ab: AB = { a: 'a', b: 1 }
+
+const abs: ReadonlyArray<AB> = [{ a: 'a', b: 1 }]
+
+_.stringify({ a: 'a', b: 1 })
+_.stringify(ab)
+_.stringify([{ a: 'a', b: 1 }])
+_.stringify(abs)